Fix import path for entity extractor
