+++ integrations/presidio/pydoc/config_docusaurus.yml
@@ -1,8 +1,8 @@
 loaders:
   - modules:
+      - haystack_integrations.components.extractors.presidio.presidio_entity_extractor
       - haystack_integrations.components.preprocessors.presidio.presidio_document_cleaner
       - haystack_integrations.components.preprocessors.presidio.presidio_text_cleaner
-      - haystack_integrations.components.preprocessors.presidio.presidio_entity_extractor
     search_path: [../src]
 processors:
   - type: filter
