+++ integrations/presidio/src/haystack_integrations/components/extractors/presidio/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from haystack_integrations.components.extractors.presidio.presidio_entity_extractor import PresidioEntityExtractor
+
+__all__ = ["PresidioEntityExtractor"]
