+++ integrations/presidio/tests/test_presidio_entity_extractor.py
@@ -9,7 +9,7 @@ import pytest
 from haystack import Document
 from haystack.core.serialization import component_from_dict, component_to_dict
 
-from haystack_integrations.components.preprocessors.presidio import PresidioEntityExtractor
+from haystack_integrations.components.extractors.presidio import PresidioEntityExtractor
 
 
 class TestPresidioEntityExtractor:
@@ -30,7 +30,7 @@ class TestPresidioEntityExtractor:
         extractor = PresidioEntityExtractor(language="en", entities=["PERSON"], score_threshold=0.6)
         data = component_to_dict(extractor, "PresidioEntityExtractor")
         expected_type = (
-            "haystack_integrations.components.preprocessors.presidio.presidio_entity_extractor.PresidioEntityExtractor"
+            "haystack_integrations.components.extractors.presidio.presidio_entity_extractor.PresidioEntityExtractor"
         )
         assert data["type"] == expected_type
         assert data["init_parameters"]["entities"] == ["PERSON"]
@@ -39,7 +39,7 @@ class TestPresidioEntityExtractor:
     def test_from_dict(self):
         data = {
             "type": (
-                "haystack_integrations.components.preprocessors.presidio"
+                "haystack_integrations.components.extractors.presidio"
                 ".presidio_entity_extractor.PresidioEntityExtractor"
             ),
             "init_parameters": {"language": "en", "entities": ["EMAIL_ADDRESS"], "score_threshold": 0.5},
